chore(router-plugin): rename "code-splitter" to "router-code-splitter"
