+++ packages/router-plugin/src/index.ts
@@ -1,4 +1,4 @@
-export { unpluginRouterCodeSplitterFactory } from './code-splitter'
+export { unpluginRouterCodeSplitterFactory } from './router-code-splitter'
 export { unpluginRouterGeneratorFactory } from './router-generator'
 export { configSchema } from './config'
 export type { Config } from './config'
