+++ packages/router-plugin/src/rspack.ts
@@ -1,5 +1,5 @@
 import { createRspackPlugin } from 'unplugin'
-import { unpluginRouterCodeSplitterFactory } from './code-splitter'
+import { unpluginRouterCodeSplitterFactory } from './router-code-splitter'
 import { configSchema } from './config'
 import { unpluginRouterGeneratorFactory } from './router-generator'
 import { unpluginRouterComposedFactory } from './composed'
